Version 0.0.2.0:Build 1000 from 04.01.2018 16 40
- Реализован тест "Отчет о прохождении сеанса боевой работы ОЭД"

+++ TelemetryAnalyzerEOS/TelemetryAnalyzerEOS/FocusValueForm.cs
@@ -5,7 +5,7 @@ namespace TelemetryAnalyzerEOS
 {
     public partial class FocusValueForm : Form
     {
-        public int FocusValue;
+        public double FocusValue;
         public FocusValueForm()
         {
             InitializeComponent();
@@ -13,9 +13,9 @@ namespace TelemetryAnalyzerEOS
 
         private void btnFocusValue_Click(object sender, EventArgs e)
         {
-            var isValid = int.TryParse(tbFocusValue.Text, out FocusValue);
-            if (isValid)
-                Hide();
+            var isValid = double.TryParse(tbFocusValue.Text, out FocusValue);
+            if (isValid)         
+               Hide();
             else
                 MessageBox.Show("Недопустимое значение фокуса, пожалуйста введите другое значение.");
         }
